Player Stopfunction fix
test angepasst

refactor
